move story  context to its own llm types
