+++ src/environment/storyteller.py
@@ -1,25 +1,8 @@
 from abc import abstractmethod
-from dataclasses import dataclass
 import re
-from typing import List, Optional
 
 from llm.llama import Llama
-
-@dataclass
-class StoryContext:
-    """
-    Represents a container for a context needed to generate the description of a location.
-    """
-    
-    """
-    Describes the general atmosphere of a story.
-    """
-    atmosphere: List[str]
-    
-    """
-    When something goes wrong e.g. player tries to move in a forbidden direction, this contains the information.
-    """
-    error: Optional[str]
+from llm.types import StoryContext
 
 class Storyteller:
     """
